requirements/base.txt
Django==1.11
pytz==2017.2
wheel==0.24.0
django-grappelli==2.9.1
django-cors-headers==2.0.2
django-cors-middleware==1.3.1
graphene-django>=1.0
psycopg2==2.7.1
django-graph-auth
django-filter==1.0.2
graphene[django]
django_graphiql